linkedin_agent/requirements.txt
linkedin-api>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.1
pandas>=2.0.0
google-generativeai>=0.8.0 